move ghquery class to utils
